requirements.txt
numpy>=1.26.4
opencv-python>=4.10.0.82
streamlit>=1.35.0
tensorflow>=2.15.0
keras>=2.15.0
annoy>=1.17.3
python-dotenv>=1.0.1
gdown>=5.2.0
tqdm>=4.66.4
st-annotated-text>=4.0.1